Included EFDatastore and improved the validation results logic
