+++ DX.Data.Xpo.Identity/DX.Data.Xpo.Identity.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <Version>20.1.2.9-pre-20064</Version>
-    <FileVersion>20.1.2.9-pre-20064</FileVersion>
+    <Version>20.1.3.10</Version>
+    <FileVersion>20.1.3.10</FileVersion>
     <Authors>Don Wibier (DevExpress)</Authors>
     <Description>XPO Based storage provider for ASP.NET Identity</Description>
     <Copyright>Copyright (c) 2017 Don Wibier</Copyright>
@@ -15,6 +15,7 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageTags>DXWeb Wibier DevExpress XPO ORM Microsoft Identity Advantage SQL Anywhere Sybase DB2 Interbase/Firebird MS-Access Oracle Pervasive PostgreSQL SQLite VistaDB MS-SQL MySQL</PackageTags>
     <PackageReleaseNotes>
+		20.1.3.10: Upgrade to DX v20.1.3
 		20.1.2.9-pre-20064: Upgrade to DX v20.1.2-pre-20064 and .NET preview 5
 		19.2.6.9: DI Extensionmethod now uses an options object for initialisation. (Recommended way)
 		Other refactorings done as well with this
@@ -60,7 +61,7 @@
 	</ItemGroup>-->
 
 	<ItemGroup>
-    <PackageReference Include="DevExpress.Xpo" Version="20.1.2-pre-20064" />
+    <PackageReference Include="DevExpress.Xpo" Version="20.1.3" />
     <!--<PackageReference Include="Microsoft.AspNet.Identity.Core" Version="2.2.2" />-->    
   </ItemGroup>
