+++ DX.Data.Xpo.Identity/XpoDiagram.xpo
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<dataBaseObjectModel xmlns:dm0="http://schemas.microsoft.com/VisualStudio/2008/DslTools/Core" dslVersion="1.0.0.0" Id="faccfc9d-b8d3-423a-b210-08be3169c70b" namespace="DX.Data.Xpo.Identity.Persistent" dataBaseShemaInfo="H4sIAAAAAAAEAO1bW28URxN9j5T/YPFMeft+iRZHxIYEhSAEOOK1uqsar7IXay825Nd/NWtjMIYvRIx3ROQXy9PTo52zVXXqnJrZ8c9vZ9O9M16uJov5g3t6X937+eDHH8a/4IpfrhdLfMNP5m2xJ7vmq5/eriYP7p2s16c/jUbn5+f753Z/sXwzMkrp0es/nr6sJzxDmMxXa5xXvnd1Ff3zVffkU/f2xq+wTHnVfeT2+P3KRwvvl57hjA+OXr9YyP7x6MPS1a5H827tYL3c8Hh0eXB18miy6haOTwnXfNBwupJN1xev9v66mdATOlCtaJeShVAygSvWAqZqoKkYbPIhqVrGo8vNVxcfLqab2Xx1DcAXl+VEh+Hi5MGTo/Hoo8Nr2/4fuH8L8GOQpJJqGBiMyg2cEXzouYLRrIN1KmvMN0DK5aNvwNn9u3ukOnhCpwJ4Enwu6wCoixfghr0yWiXnbwPp8ekpL3cPV7LX6eYrVOcNOF0VJO9QUlj7oFRwQdvbgPtssZzhdPI30+4xV+OtUcmAZuXAtZIgOcqAXDQRWcPW9I75Cu8wad0r5s+tXnLtv6Do45W0l51QdGxKFaEoyCVIkhMZKJa0lLivkT21lL5/ijbBUq3ZQTJBgWMnlcwhArF1HrNuuX+KfjTDyXT3UAtlaawmAyuWbhSihRwwCmm10lS1mEq8FaiHi3mbLGdDkJYrgQI1BYWNlg6MBZK1gp69zg2LLrF30nqOq9X5Ykm/4epkgE5MSNFJLicrAXZGYJfcGCTNQ84S+oiqb8QvuW6Wk/W7l2ucnQ7A0iFmY1URosoiPqhEIapcoBL70qSOm71JVN8a5JPFnJ9tZmUI+WFy0o4wAClJaCdRFcDNQyJKtemadWm3CHjAejY5RooS28BNGrKlIshNAsxRWR/QGds7h706XzzGKvbtAscAqFP0OXqrIZsQOrPURG76BlFhTIiaQutdXT9d1L8Wm/WjOR3JbR2v6xDqK1jLLUHNQdI8OOnSLimoxVYbuZG2fGuwBwp1ramIK3TC1wGlTQncQpxA+WhscQVD7T3UD2vl1eqxdGqmw8Vmvh6ga1XfxCwnUSRNQk3KQmqOoBpUkvqBWae+YXdKej6IrYhZ1xJVx99GWExF6dG2WODkExrWLlp3G2iHMVFDoN3qzoGGA9Yb05LWgF7UtWvad/8RtOjJ6uS0y70PB95HdyDIldH5oAxYVt2gKyqxyMLZXpeQVcVgW+9MvQ3xkAMR35C0gIOQSKxj6gRYRIako7RppYN2txbnQQdBOqE2ojm1IpK2TFXUiOD21orlMNFp1Lc3CBrIPA8R7A+ghyPvYYP9H8Dc39Tv6eLNZL6jpzNRDFUSFWpq6/i8ZVHe3KB62zAEUd+fUd7fMvrbYnu+XJxNaIj+Rc5mh4RglRPVjU0Ku4QmOiU0E3wquSPang31Jdrf+d0As0AdxWBUhuJ1BedjZzOKcBqLoyQspbjefXSXwwM4C1O6Ea+YimqyOItooCBrEM62PlSXE329oeqvmA+nOJntpphLiaVFq8C7Il8DKrHR2jE0U0zVwcZUbk4Ev7c5PlWdqxLxaauSCrZMkFJukGorWnnjw2dAfpcJLQScCnsDOpgALmqEwuwhI7ZMZIw3vT+x2Gbrq3enQ7hHxMC1IVDORopYShdrIDChxepjTdH2LkC2cP/E6WYIvFRas8ZAajkJX3mCIhITSjJJe2sK+q9P5D746vXp4uht9wbLTshKXKMpvgpJe3LSmFyA7ESDBK9KSNU7VW8+sPkmshrAQWElRdKChayigDQZRW5oDVVTLilIS6be7fJDou3jmeNXhwO4p2Q1YooiphODxNBLE3YdZ2mdNcdCqfcR0B+LAQFLtZIIRwsUHF1UMeYkUppsUV7JnxJ2X8Vdw3q1+IvnOynl7E0jrmKeqAryZiIUZRRE1Apt1aa1ft8fGKYf1yDySjQzkLJe+jGJbfDGQdGeEFHpoG7h4cygdimlGhMnB52w7Oxh9zwqK6jZq6aUpphv5W2vAYgr2qxtsqIjJcLOe9epjwDNYuk4nLLuXVYOJDykHkMNPkOo1L3uIzUrLlAB+UamRq2j//pHML1S1laO7YSyJKrZJCljL5ClT7kKWamOt+RLsU7y/jON+fujLKdrxVQQTKc6XBCOzq1zx1YEiChszvHr3X9voe5ePt5dqJsv1vmkOjUi3alrTKlFA2yV0z6xz/rmOxTfEuoO3gB2omUfJKxQsu0CHITA0Ip59KilBfugcLfjjys7sbtQG6oYLUkZFyXxNiLGcogNVJOi5pgM9jzNHNAuq9ZcNlXkZlAWHGsFGEOBlIwNkYuz7XamAwN1rcasvWsCtaH8SSxShJAh15a8bU7C/vVWqp/x3g5He3eZfZfZO8zsqw69m+yuOeSA1DUrL84qiBBLTUdQwZbAktzSxP4DLTr4Flr1BlA8FThHFrLPDmpNUUWOtVId4AnFdlCwm0DfTQruJgV3k4K7ScGXCOvy+MOWcfdrWibJ67pd3BtdLP958dPbgyM+e/T2dMmr1f4RrnH/TOd9fX/v8vSD7mjf7av7e4eb6Xqz5Adz3qyXOL2/93xTppP6O7/b0t+DkhIJHEdRKXZZ2s37j9h+njTD1YpXXQBebK6+3fFDklvjNnl7eV8XSy837drSC54tzvjTjRern+49FMN0cDyvJzh/072KsD2++Go+dw9j+WZOebmeDHxvX7iN8a8856Wk1+FiNuP5enWZozeWr20+4tXkjfz7eCKp8MkV189tL3u2mU67rPmFT/BsslhKUjSUcEu9fHrmxx/Go09+o33wP3gmaEPMPQAA" generateStoredProcedures="false" fieldNameCase="Unchanged" fieldNamePrefix="_" generateDefaultConstructors="false" xmlns="http://schemas.microsoft.com/dsltools/ORMDataModel">
+<dataBaseObjectModel xmlns:dm0="http://schemas.microsoft.com/VisualStudio/2008/DslTools/Core" dslVersion="1.0.0.0" Id="faccfc9d-b8d3-423a-b210-08be3169c70b" namespace="DX.Data.Xpo.Identity.Persistent" dataBaseShemaInfo="H4sIAAAAAAAEAO1bW28URxN9j5T/YPFMeft+iRZHxIYEhSAEOOK1uqsar7IXay825Nd/NWtjMIYvRIx3ROQXy9PTo52zVXXqnJrZ8c9vZ9O9M16uJov5g3t6X937+eDHH8a/4IpfrhdLfMNP5m2xJ7vmq5/eriYP7p2s16c/jUbn5+f753Z/sXwzMkrp0es/nr6sJzxDmMxXa5xXvnd1Ff3zVffkU/f2xq+wTHnVfeT2+P3KRwvvl57hjA+OXr9YyP7x6MPS1a5H827tYL3c8Hh0eXB18miy6haOTwnXfNBwupJN1xev9v66mdATOlCtaJeShVAygSvWAqZqoKkYbPIhqVrGo8vNVxcfLqab2Xx1DcAXl+VEh+Hi5MGTo/Hoo8Nr2/4fuH8L8GOQpJJqGBiMyg2cEXzouYLRrIN1KmvMN0DK5aNvwNn9u3ukOnhCpwJ4Enwu6wCoixfghr0yWiXnbwPp8ekpL3cPV7LX6eYrVOcNOF0VJO9QUlj7oFRwQdvbgPtssZzhdPI30+4xV+OtUcmAZuXAtZIgOcqAXDQRWcPW9I75Cu8wad0r5s+tXnLtv6Do45W0l51QdGxKFaEoyCVIkhMZKJa0lLivkT21lL5/ijbBUq3ZQTJBgWMnlcwhArF1HrNuuX+KfjTDyXT3UAtlaawmAyuWbhSihRwwCmm10lS1mEq8FaiHi3mbLGdDkJYrgQI1BYWNlg6MBZK1gp69zg2LLrF30nqOq9X5Ykm/4epkgE5MSNFJLicrAXZGYJfcGCTNQ84S+oiqb8QvuW6Wk/W7l2ucnQ7A0iFmY1URosoiPqhEIapcoBL70qSOm71JVN8a5JPFnJ9tZmUI+WFy0o4wAClJaCdRFcDNQyJKtemadWm3CHjAejY5RooS28BNGrKlIshNAsxRWR/QGds7h706XzzGKvbtAscAqFP0OXqrIZsQOrPURG76BlFhTIiaQutdXT9d1L8Wm/WjOR3JbR2v6xDqK1jLLUHNQdI8OOnSLimoxVYbuZG2fGuwBwp1ramIK3TC1wGlTQncQpxA+WhscQVD7T3UD2vl1eqxdGqmw8Vmvh6ga1XfxCwnUSRNQk3KQmqOoBpUkvqBWae+YXdKej6IrYhZ1xJVx99GWExF6dG2WODkExrWLlp3G2iHMVFDoN3qzoGGA9Yb05LWgF7UtWvad/8RtOjJ6uS0y70PB95HdyDIldH5oAxYVt2gKyqxyMLZXpeQVcVgW+9MvQ3xkAMR35C0gIOQSKxj6gRYRIako7RppYN2txbnQQdBOqE2ojm1IpK2TFXUiOD21orlMNFp1Lc3CBrIPA8R7A+ghyPvYYP9H8Dc39Tv6eLNZL6jpzNRDFUSFWpq6/i8ZVHe3KB62zAEUd+fUd7fMvrbYnu+XJxNaIj+Rc5mh4RglRPVjU0Ku4QmOiU0E3wquSPang31Jdrf+d0As0AdxWBUhuJ1BedjZzOKcBqLoyQspbjefXSXwwM4C1O6Ea+YimqyOItooCBrEM62PlSXE329oeqvmA+nOJntpphLiaVFq8C7Il8DKrHR2jE0U0zVwcZUbk4Ev7c5PlWdqxLxaauSCrZMkFJukGorWnnjw2dAfpcJLQScCnsDOpgALmqEwuwhI7ZMZIw3vT+x2Gbrq3enQ7hHxMC1IVDORopYShdrIDChxepjTdH2LkC2cP/E6WYIvFRas8ZAajkJX3mCIhITSjJJe2sK+q9P5D746vXp4uht9wbLTshKXKMpvgpJe3LSmFyA7ESDBK9KSNU7VW8+sPkmshrAQWElRdKChayigDQZRW5oDVVTLilIS6be7fJDou3jmeNXhwO4p2Q1YooiphODxNBLE3YdZ2mdNcdCqfcR0B+LAQFLtZIIRwsUHF1UMeYkUppsUV7JnxJ2X8Vdw3q1+IvnOynl7E0jrmKeqAryZiIUZRRE1Apt1aa1ft8fGKYf1yDySjQzkLJe+jGJbfDGQdGeEFHpoG7h4cygdimlGhMnB52w7Oxh9zwqK6jZq6aUpphv5W2vAYgr2qxtsqIjJcLOe9epjwDNYuk4nLLuXVYOJDykHkMNPkOo1L3uIzUrLlAB+UamRq2j//pHML1S1laO7YSyJKrZJCljL5ClT7kKWamOt+RLsU7y/jON+fujLKdrxVQQTKc6XBCOzq1zx1YEiChszvHr3X9voe5ePt5dqJsv1vmkOjUi3alrTKlFA2yV0z6xz/rmOxTfEuoO3gB2omUfJKxQsu0CHITA0Ip59KilBfugcLfjjys7sbtQG6oYLUkZFyXxNiLGcogNVJOi5pgM9jzNHNAuq9ZcNlXkZlAWHGsFGEOBlIwNkYuz7XamAwN1rcasvWsCtaH8SSxShJAh15a8bU7C/vVWqp/x3g5He3eZfZfZO8zsqw69m+yuOeSA1DUrL84qiBBLTUdQwZbAktzSxP4DLTr4Flr1BlA8FThHFrLPDmpNUUWOtVId4AnFdlCwm0DfTQruJgV3k4K7ScGXCOvy+MOWcfdrWibJ67pd3BtdLP958dPbgyM+e/T2dMmr1f4RrnH/zKh9fX/v8vSD7mjf7qv7e4eb6Xqz5Adz3qyXOL2/93xTppP6O7/b0t+DkhIJHEdRKXZZ2s37j9h+njTD1YpXXQBebK6+3fFDklvjNnl7eV8XSy837drSC54tzvjTjRern+49FMN0cDyvJzh/072KsD2++Go+dw9j+WZOebmeDHxvX7iN8a8856Wk1+FiNuP5enWZozeWr20+4tXkjfz7eCKp8MkV189tL3u2mU67rPmFT/BsslhKUjSUcEu9fHrmxx/Go09+o33wP31p2GrMPQAA" generateStoredProcedures="false" fieldNameCase="Unchanged" fieldNamePrefix="_" generateDefaultConstructors="false" generateJsonSerializationContractResolver="false" generateServiceCollectionExtensions="false" lastModifiedInVersion="20.1.3.0" name="XpoDiagram" xmlns="http://schemas.microsoft.com/dsltools/ORMDataModel">
   <classes>
     <dataBaseObjectModelHasClasses Id="80f46174-38a5-4740-9546-6286fe4f8797">
       <xpObject Id="0fb14883-6b9d-4b33-a8c2-f076385680cb" name="XpoDxRole" baseClass="XpoDxBase" userFileName="XpoDxRole" designerFileName="XpoDxRole.Designer" className="XpoDxRole" virtualBaseClass="XpoDxBase" persistent="DXRoles" virtualName="XpoDxRole">
