+++ DX.Data.Xpo.IndentityServer/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace DX.Data.Xpo.IndentityServer
+{
+	public class Class1
+	{
+	}
+}
