+++ DX.Data.Xpo.IndentityServer/DX.Data.Xpo.IndentityServer.csproj
@@ -1,0 +1,19 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>netstandard2.1;net461</TargetFramework>
+  </PropertyGroup>
+  <ItemGroup>
+	<PackageReference Include="IdentityServer4.Storage" Version="3.1.3" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\DX.Data.Xpo.Identity\DX.Data.Xpo.Identity.csproj" />
+    <ProjectReference Include="..\DX.Data.Xpo\DX.Data.Xpo.csproj" />
+    <ProjectReference Include="..\DX.Utils\DX.Utils.csproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Update="Diagram.xpo.diagram">
+      <AutoGen>True</AutoGen>
+    </None>
+  </ItemGroup>
+</Project>
