+++ DX.Data.Xpo.Mvc/DX.Data.Xpo.Mvc.csproj
@@ -42,8 +42,8 @@
 	</ItemGroup>
 	
 	<ItemGroup>
-		<PackageReference Include="DevExpress.Xpo" Version="20.1.2-pre-20064" />
-		<PackageReference Include="DevExpress.Web.Mvc5" Version="20.1.2-pre-20064" />
+		<PackageReference Include="DevExpress.Xpo" Version="20.1.3" />
+		<PackageReference Include="DevExpress.Web.Mvc5" Version="20.1.3" />
 	</ItemGroup>
 
 	<ItemGroup>
