+++ DX.Data.Xpo/DX.Data.Xpo.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<Version>20.1.2.13-pre-20064</Version>
-		<FileVersion>20.1.2.13-pre-20064</FileVersion>
+		<Version>20.1.3.14</Version>
+		<FileVersion>20.1.3.14</FileVersion>
 		<Authors>Don Wibier (DevExpress)</Authors>
 		<Description>Several XPO Related Helper classes</Description>
 		<Copyright>Copyright (c) 2017 Don Wibier</Copyright>
@@ -15,6 +15,7 @@
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<PackageTags>DXWeb Wibier DevExpress XPO ORM Microsoft Identity Advantage SQL Anywhere Sybase DB2 Interbase/Firebird MS-Access Oracle Pervasive PostgreSQL SQLite VistaDB MS-SQL MySQL</PackageTags>
 		<PackageReleaseNotes>
+			20.1.3.14: Upgrade to DX v20.1.3
 			20.1.2.13-pre-20064: Upgrade to DX v20.1.2-pre-20064 and .NET preview 5
 			19.2.6.13: Bug fixed in overloaded method to get XpoDatabase instance with options object
 			19.2.6.12: DI Extensionmethod now uses an options object for initialisation. Other refactorings done as well
@@ -57,12 +58,13 @@
 	<ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
 		<PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
 
-		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="5.0.0-preview.2.20160.3" />
-		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0-preview.2.20160.3" />
+		<PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="5.0.0-preview.3.20215.2" />
+		<PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="5.0.0-preview.3.20215.2" />
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="DevExpress.Xpo" Version="20.1.2-pre-20064" />
+		<PackageReference Include="DevExpress.Xpo" Version="20.1.3" />
+		<!--<PackageReference Include="DevExpress.Data" Version="20.1.3" />-->
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\DX.Utils\DX.Utils.csproj" />
