+++ DX.Test.Web.Core/DX.Test.Web.Core.csproj
@@ -7,12 +7,12 @@
 
 
   <ItemGroup>
-    <PackageReference Include="DevExpress.Xpo" Version="20.1.2-pre-20064" />
+    <PackageReference Include="DevExpress.Xpo" Version="20.1.3" />
 	<!--<PackageReference Include="Microsoft.AspNetCore.Identity" Version="3.1.0" />-->
-	<PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.0-preview.2.20167.3" />
+	<PackageReference Include="Microsoft.AspNetCore.Identity.UI" Version="5.0.0-preview.3.20215.14" />
 
-    <PackageReference Include="Microsoft.Data.Sqlite" Version="5.0.0-preview.2.20159.4" />
-    <PackageReference Include="Microsoft.Data.Sqlite.Core" Version="5.0.0-preview.2.20159.4" />
+    <PackageReference Include="Microsoft.Data.Sqlite" Version="5.0.0-preview.3.20181.2" />
+    <PackageReference Include="Microsoft.Data.Sqlite.Core" Version="5.0.0-preview.3.20181.2" />
 
 	  <!--<PackageReference Include="System.Data.SQLite.Core" Version="1.0.112" />-->
   </ItemGroup>
