+++ DX.Test.Web.MVC/packages.config
@@ -2,17 +2,17 @@
 <packages>
   <package id="Antlr" version="3.5.0.2" targetFramework="net461" />
   <package id="bootstrap" version="4.4.1" targetFramework="net461" />
-  <package id="DevExpress.Data" version="20.1.2-pre-20064" targetFramework="net461" />
-  <package id="DevExpress.Xpo" version="20.1.2-pre-20064" targetFramework="net461" />
-  <package id="jQuery" version="3.5.0.1" targetFramework="net461" />
+  <package id="DevExpress.Data" version="20.1.3" targetFramework="net461" />
+  <package id="DevExpress.Xpo" version="20.1.3" targetFramework="net461" />
+  <package id="jQuery" version="3.5.1" targetFramework="net461" />
   <package id="jQuery.Validation" version="1.19.1" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights" version="2.14.0-beta5" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights" version="2.15.0-beta1" targetFramework="net461" />
   <package id="Microsoft.ApplicationInsights.Agent.Intercept" version="2.4.0" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.14.0-beta5" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.14.0-beta5" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights.Web" version="2.14.0-beta5" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.14.0-beta5" targetFramework="net461" />
-  <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.14.0-beta5" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights.DependencyCollector" version="2.15.0-beta1" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights.PerfCounterCollector" version="2.15.0-beta1" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights.Web" version="2.15.0-beta1" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights.WindowsServer" version="2.15.0-beta1" targetFramework="net461" />
+  <package id="Microsoft.ApplicationInsights.WindowsServer.TelemetryChannel" version="2.15.0-beta1" targetFramework="net461" />
   <package id="Microsoft.AspNet.Identity.Core" version="2.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.Identity.Owin" version="2.2.3" targetFramework="net461" />
   <package id="Microsoft.AspNet.Mvc" version="5.2.7" targetFramework="net461" />
@@ -22,7 +22,7 @@
   <package id="Microsoft.AspNet.WebPages" version="3.2.7" targetFramework="net461" />
   <package id="Microsoft.CodeDom.Providers.DotNetCompilerPlatform" version="2.0.1" targetFramework="net461" />
   <package id="Microsoft.jQuery.Unobtrusive.Validation" version="3.2.11" targetFramework="net461" />
-  <package id="Microsoft.Net.Compilers" version="3.6.0-3.final" targetFramework="net461" developmentDependency="true" />
+  <package id="Microsoft.Net.Compilers" version="3.6.0-4.final" targetFramework="net461" developmentDependency="true" />
   <package id="Microsoft.Owin" version="4.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Host.SystemWeb" version="4.1.0" targetFramework="net461" />
   <package id="Microsoft.Owin.Security" version="4.1.0" targetFramework="net461" />
@@ -38,9 +38,9 @@
   <package id="Owin" version="1.0" targetFramework="net461" />
   <package id="popper.js" version="1.16.1" targetFramework="net461" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net461" />
-  <package id="System.Diagnostics.DiagnosticSource" version="5.0.0-preview.2.20160.6" targetFramework="net461" />
+  <package id="System.Diagnostics.DiagnosticSource" version="5.0.0-preview.3.20214.6" targetFramework="net461" />
   <package id="System.Memory" version="4.5.4" targetFramework="net461" />
   <package id="System.Numerics.Vectors" version="4.6.0-preview5.19224.8" targetFramework="net461" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="5.0.0-preview.2.20160.6" targetFramework="net461" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="5.0.0-preview.3.20214.6" targetFramework="net461" />
   <package id="WebGrease" version="1.6.0" targetFramework="net461" />
 </packages>
