+++ DX.Utils/DX.Utils.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <Version>20.1.2.14-pre-20064</Version>
-    <FileVersion>20.1.2.14-pre-20064</FileVersion>
+    <Version>20.1.3.15</Version>
+    <FileVersion>20.1.3.15</FileVersion>
     <Authors>Don Wibier (DevExpress)</Authors>
     <Description>Several C# utility classes and extension methods</Description>
     <Copyright>Copyright (c) 2019 Don Wibier</Copyright>	  
@@ -15,6 +15,7 @@
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
     <PackageTags>DXWeb Wibier DevExpress</PackageTags>
     <PackageReleaseNotes>
+		20.1.3.15: Upgrade to DX v20.1.3
 		20.1.2.14-pre-20064: Upgrade to DX v20.1.2-pre-20064 and .NET preview 5
 		1.0.0.14: BREAKING CHANGE: Signature-change on IDataValidator&lt;..&gt;.Deleting(..):
 			        Was: Deleting(TKey id, object arg, IDataValidationResults&lt;TKey&gt; validationResults);
@@ -43,9 +44,7 @@
   </ItemGroup>
 
   <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
-    <PackageReference Include="System.Configuration.ConfigurationManager">
-      <Version>5.0.0-preview.2.20160.6</Version>
-    </PackageReference>
+    <PackageReference Include="System.Configuration.ConfigurationManager" Version="4.7.0" />
   </ItemGroup>
 
   <Target Name="PostBuild" AfterTargets="PostBuildEvent">
