+++ DX.Utils/Data/DataStore.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Configuration;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -8,18 +9,24 @@ namespace DX.Utils.Data
 {
 	public abstract class DataStore<TKey, TModel> : IDataStore<TKey, TModel>, IDisposable
 		where TKey : IEquatable<TKey>
-		where TModel : IDataStoreModel<TKey>		
+		where TModel : IDataStoreModel<TKey>
 	{
 		public virtual Type KeyType => typeof(TKey);
 		public virtual Type ModelType => typeof(TModel);
 
 		public abstract TModel GetByKey(TKey key);
+
+		//public abstract IDataMapper<TKey, TModel, TDBModel> GetMapper<TDBModel>()
+		//	where TDBModel : class, IDataStoreModel<TKey>;
+
+		//public abstract IDataStoreValidator<TKey, TModel> GetValidator();
+
 		public async virtual Task<TModel> GetByKeyAsync(TKey key)
 		{
 			var result = await Task.FromResult(GetByKey(key));
 			return result;
 		}
-		protected abstract IEnumerable<TModel> Query();		
+		protected abstract IEnumerable<TModel> Query();
 
 		//public async virtual Task<IEnumerable<TModel>> QueryAsync()
 		//{
@@ -38,7 +45,7 @@ namespace DX.Utils.Data
 		{
 			var result = await Task.FromResult(Create(items));
 			return result;
-			
+
 		}
 		public async virtual Task<IDataValidationResults<TKey>> CreateAsync(TModel item)
 		{
@@ -57,18 +64,54 @@ namespace DX.Utils.Data
 		public async virtual Task<IDataValidationResults<TKey>> UpdateAsync(IEnumerable<TModel> items)
 		{
 			var result = await Task.FromResult(Update(items));
-			return result;			
+			return result;
 		}
 		public async virtual Task<IDataValidationResults<TKey>> UpdateAsync(TKey id, TModel item)
 		{
 			var result = await Task.FromResult(Update(id, item));
-			return result;			
+			return result;
 		}
 		public async virtual Task<IDataValidationResults<TKey>> UpdateAsync(TModel item)
 		{
 			var result = await Task.FromResult(Update(item));
-			return result;			
+			return result;
 		}
+
+		protected TKey EmptyKeyValue => default;
+
+		/// <summary>
+		/// When ID = null/default, it will add otherwise it will update and collect the results
+		/// </summary>
+		/// <param name="items"></param>
+		/// <returns></returns>
+		public virtual IDataValidationResults<TKey> Store(IEnumerable<TModel> items)
+		{
+			var updated = Update(items.Where(o => !o.ID.Equals(EmptyKeyValue)));
+			var inserted = Create(items.Where(o => o.ID.Equals(EmptyKeyValue)));
+			foreach (var r in inserted.Results)
+				updated.Add(r);
+			return updated;
+		}
+
+		public virtual IDataValidationResults<TKey> Store(TKey id, TModel item)
+		{
+			var result = id.Equals(EmptyKeyValue) ? Create(item) : Update(id, item);
+			return result;
+		}
+
+		public async virtual Task<IDataValidationResults<TKey>> StoreAsync(IEnumerable<TModel> items)
+		{
+			var result = await Task.FromResult(Store(items));
+			return result;
+		}
+
+		public async virtual Task<IDataValidationResults<TKey>> StoreAsync(TKey id, TModel item)
+		{
+			var result = await Task.FromResult(Store(id, item));
+			return result;
+		}
+
+
 		public abstract IDataValidationResults<TKey> Delete(IEnumerable<TKey> ids);
 		public virtual IDataValidationResults<TKey> Delete(IEnumerable<TModel> items)
 		{
@@ -118,7 +161,7 @@ namespace DX.Utils.Data
 				if (disposing)
 				{
 					// TODO: dispose managed state (managed objects).
-					
+
 				}
 
 				// TODO: free unmanaged resources (unmanaged objects) and override a finalizer below.
@@ -142,6 +185,7 @@ namespace DX.Utils.Data
 			// TODO: uncomment the following line if the finalizer is overridden above.
 			// GC.SuppressFinalize(this);
 		}
+
 		#endregion
 
 	}
