+++ DXWeb.sln
@@ -33,7 +33,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Test Projects", "Test Proje
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DX.Test.Web.Blazor", "DX.Test.Web.Blazor\DX.Test.Web.Blazor.csproj", "{75813F2B-496F-44D3-80D6-4943C7D8822B}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DX.Data.EF", "DX.Data.EF\DX.Data.EF.csproj", "{26AF7CF2-E0F2-49F7-B3AD-622B2957DE49}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DX.Data.EF", "DX.Data.EF\DX.Data.EF.csproj", "{26AF7CF2-E0F2-49F7-B3AD-622B2957DE49}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DX.Data.Xpo.IndentityServer", "DX.Data.Xpo.IndentityServer\DX.Data.Xpo.IndentityServer.csproj", "{AD3CA4B3-8E9F-45A0-9477-DB8CA7E83792}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -69,6 +71,10 @@ Global
 		{26AF7CF2-E0F2-49F7-B3AD-622B2957DE49}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{26AF7CF2-E0F2-49F7-B3AD-622B2957DE49}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{26AF7CF2-E0F2-49F7-B3AD-622B2957DE49}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AD3CA4B3-8E9F-45A0-9477-DB8CA7E83792}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AD3CA4B3-8E9F-45A0-9477-DB8CA7E83792}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AD3CA4B3-8E9F-45A0-9477-DB8CA7E83792}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AD3CA4B3-8E9F-45A0-9477-DB8CA7E83792}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
